# Optimise 'Math.CopySign' and 'MathF.CopySign'