--- a/src/System.Private.CoreLib/shared/System/MathF.cs
+++ b/src/System.Private.CoreLib/shared/System/MathF.cs
@@ -95,7 +95,7 @@ namespace System
         {
             const int signMask = unchecked((int)0b_1000_0000__0000_0000__0000_0000__0000_0000);
 
-            if (Sse.IsSupported)
+            if (Sse.X64.IsSupported)
             {
                 var xvec = Vector128.CreateScalarUnsafe(x);
                 var yvec = Vector128.CreateScalarUnsafe(y);