--- a/src/System.Private.CoreLib/shared/System/MathF.cs
+++ b/src/System.Private.CoreLib/shared/System/MathF.cs
@@ -103,8 +103,8 @@ namespace System
                 // Remove the sign from x, and remove everything but the sign from y
                 // Creating from a 'const long' is better than from the correct 'const float/double'
                 var mask = Vector128.CreateScalarUnsafe(signMask).AsSingle();
+                yvec = Sse.And(yvec, mask);
                 xvec = Sse.AndNot(mask, xvec);
-                yvec = Sse.And(mask, yvec);
 
                 // Simply OR them to get the correct sign
                 return Sse.Or(xvec, yvec).ToScalar();