@john-h-k 2019-09-04
@danmosemsft 

> Well @john-h-k perhaps you might consider adding a benchmark (such as some form of the one you're using here) to https://github.com/dotnet/performance as well ? There's excellent documentation there: it's easy to use. That would protect your change.

Looking at that right now - will make sure to do so