@mikedn 2019-09-04
> The SSE pathway doesn't spill to the stack, unlike the others

The spill that happens on the non-SSE path is something that can be fixed in the JIT. I'm not sure if once that is done the SSE path would still be significantly faster to warrant the extra complexity as the ANDPS/ORPS instructions have lower throughput compared to the their integer counterparts.

For benchmarking purposes it would be useful to test the non-SSE version with `SingleToInt64Bits` replaced by SSE's MOVD intrinsic.