@mikedn 2019-09-04
> For benchmarking purposes it would be useful to test the non-SSE version with SingleToInt64Bits replaced by SSE's MOVD intrinsic.

Though with `ANDNPS` I suspect that there's no way that the non-SSE version can be faster, even with `MOVD`. The SSE code with `ANDNPS` is:
```asm
       48B80000000000000080 mov      rax, 0x8000000000000000
       C4E1F96ED0           vmovd    xmm2, rax
       C5E854C9             vandps   xmm1, xmm2, xmm1
       C5E855C0             vandnps  xmm0, xmm2, xmm0
       C5F856C1             vorps    xmm0, xmm0, xmm1
```
which is probably as good as it gets.