@EgorBo 2019-09-04
just for reference here is what mono generates for:
```csharp
 [MethodImpl(MethodImplOptions.NoInlining)]
 public static double Test(double x, double y) => Math.CopySign(x, y);
```
asm:
```asm
0000000000000000	movabsq	$0x7fb20121f3e0, %rax
000000000000000a	vandps	(%rax), %xmm1, %xmm1
000000000000000e	movabsq	$0x7fb20121f3f0, %rax
0000000000000018	vandps	(%rax), %xmm0, %xmm0
000000000000001c	vorps	%xmm1, %xmm0, %xmm0
0000000000000020	retq
```