@tannergooding 2019-09-04
> you might consider adding a benchmark

That would be greatly appreciated and would make validating this change on other boxes simpler 😄. In general, the benchmarks are being checked into the dotnet/performance repo. -- @adamsitnik, do we have any guidance on how to run dotnet/performance benchmarks against local dotnet/coreclr builds?

Overall the changes LGTM, just want to test this on a few more boxes (and probably an ARM machine) before committing it.