@john-h-k 2019-09-04
Testing on [sharplab](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgEYBYAKGIAYACY8gOgCUBXAOwwEt8YLAMIR8AB14AbGFADKMgG68wMXAG4a9Jq049+ggJI8ovLrmXrNjZu258BLIxhNmLLABoAOJBuqaAzEykDEIMAN40DFEMAPQxANoAsjAYABYQACYG4pIAFMlpmdlikgDyYnwQZiwAggDmdbC45gowRpKmpnUAlAC6kdHEgcxIDNy42ABmMAwZEBzA0iEQYgCesrx1XLlzC0sIaLPzizOr3QNREdTRNwyQZhgMklV1DOZbiXgA1gwXtzcAXjGXDAqRgYC+MAyuVyzy4PTowAA+uQ6GikWj0RjMdisZi6LiCYTicT8UjSTiKXicfjur4/jdeJMGLlZLhDLhZBwxGJoBgoedrv9wgzhXEQrBsPyGK0wBhoLgGBBmWkZjBpAIeLhDtguBkGPhsJCGKr8AxJs8pSaIAxsJJngB3JiMWSyACiDCgUrBUBNqV1DFdbuCplwzg4ct4VVwov+CmwvoQsoYQIAauD5VByKRPMJJfzZGA7QmAKpmKYwXIIbq1TldaS5Omx27x32rZNpjPQbO5oT5mCF4tQMsTaa5M61jbwhtNvxC4XRcVsGD4CCtP0zd5cc1QUQMA62vWeldrtWtKCrNJdBjADiPVVvTbbya7s2rZs3cV9mBS68vvfYAwADk9xhk8LxAQwvCKsAKT8r6aQBv+ZoPpAUCwHKwGgY8FoQFKMS7CcQEfouMStga3wpgw6Zyt2OZ5j+BZFpIpblmOW6fLgXw1jUdbTpWs4Ljc5GGlxVE0ZmPYMVKA7Maxo6VnAdAsHQkw8XxdQzr4QnREm4JUWygg1HqAByEAYLkolfIcelgIJOkMO2+lAoZtR6pZ3yHE5dn0vOQnihsOSrAwpRsBuKE2nUKQbnc0AYY8W4kVExAAOyBuyLClFAVayl5srqQAIscM4sAAKhAg4sdl9n/AAvs26rss2VwObEMQMKVqTQQaKTpPq3WwAAjhwvCwPq8oMA60A/JM0C2vaUFcGId7aklbWLWAkgcBk14mdgJmHLgNoOjMKwyDJSrbg+XpOsAvAYLgLA0GtcKvAgd0PVRABC90iFw57wSwRV7DA5VOKgP0PVWNVCa9jkfYqQKQ39AMyEDxWgxA4MoJDuDjrOa1Lie64PluO57geur6rAq7rjA56Xl18I3neMVk8hjnPX5C7vfdioAGRAgAfhx3zaQ5qwIwwguPh8Ytc61AX8CUwWheF1oMFF95grF6EZrLXBralDDI1UqNQI4PAQ3z5XAycVZS4Q8N8zDNz1X57s3H8QzaKM4wVkcIPLGsGxbEiOwY/uhyEUsZx/C1wrGyIIdPlWXkw+7tVAA=), the JIT doesn't seem to inline this method in its current state, but will with `aggressiveinlining` applied. Is it worth applying it to the method given the discussion around its use in tight loops 
> which I believe could be impactful to tight loops.