@mikedn 2019-09-04
The SSE `double` version is problematic on x86:
```asm
       C5FB10442444 vmovsd   xmm0, qword ptr [esp+44H]
       C5F911442420 vmovupd  xmmword ptr [esp+20H], xmm0
       C5FB104C243C vmovsd   xmm1, qword ptr [esp+3CH]
       C5F9114C2410 vmovupd  xmmword ptr [esp+10H], xmm1
       6800000080   push     0x80000000
       6A00         push     0
       8D4C2408     lea      ecx, bword ptr [esp+08H]
       FF1514C08E05 call     [System.Runtime.Intrinsics.Vector128:CreateScalarUnsafe(long):struct]
       C5F9100424   vmovupd  xmm0, xmmword ptr [esp]
       C5F855442420 vandnps  xmm0, xmm0, xmmword ptr [esp+20H]
       C5F9104C2410 vmovupd  xmm1, xmmword ptr [esp+10H]
       C5F0540C24   vandps   xmm1, xmm1, xmmword ptr [esp]
       C5F856C1     vorps    xmm0, xmm0, xmm1
       C5FB11442430 vmovsd   qword ptr [esp+30H], xmm0
       DD442430     fld      qword ptr [esp+30H]
```
It probably should only be enabled on x64.