@tannergooding 2019-09-04
> It probably should only be enabled on x64.

Sounds good to me.

It's worth noting that some of that looks like codegen issues we might want to track... There isn't really any reason why this couldn't be:
```asm
vmovq  xmm2, qword ptr [address of ulong constant]  ; This encoding of movq is supported even on 32-bit
vandps xmm1, xmm2, xmm1
vandnps xmm0, xmm2, xmm0
vorps  xmm0, xmm0, xmm1
```