@adamsitnik 2019-09-04
> do we have any guidance on how to run dotnet/performance benchmarks against local dotnet/coreclr builds?

yes, all you need to do is to pass the path to corerun executable via `--corerun` parameter

https://github.com/dotnet/performance/tree/master/src/benchmarks/micro#private-runtime-builds

@john-h-k please let me know if something is not clear or does not work as expected